+++ Graphics_editor/src/manager_of_canvas_managers.h
@@ -75,7 +75,7 @@ class Manager_of_canvas_managers : public View_object {
 
 	void add_new_canvas_manager();
 
-	bool check_click(const double mouse_x, const double mouse_y, const Mouse_click_state* par_mouse_status) override {
+	bool check_click(const float mouse_x, const float mouse_y, const Mouse_click_state* par_mouse_status) override {
 		printf("view_manager check_click, now active %d\n", who_is_active);
 
 		//if(is_active) {
